Refactor ContinuousMLPBaseline (#1686)
