+++ src/garage/tf/regressors/__init__.py
@@ -2,8 +2,6 @@
 from garage.tf.regressors.bernoulli_mlp_regressor import BernoulliMLPRegressor
 from garage.tf.regressors.categorical_mlp_regressor import (
     CategoricalMLPRegressor)
-from garage.tf.regressors.continuous_mlp_regressor import (
-    ContinuousMLPRegressor)
 from garage.tf.regressors.gaussian_cnn_regressor_model import (
     GaussianCNNRegressorModel)
 from garage.tf.regressors.gaussian_mlp_regressor import GaussianMLPRegressor
@@ -11,6 +9,6 @@ from garage.tf.regressors.regressor import Regressor, StochasticRegressor
 
 __all__ = [
     'BernoulliMLPRegressor', 'CategoricalMLPRegressor',
-    'ContinuousMLPRegressor', 'GaussianCNNRegressorModel',
-    'GaussianMLPRegressor', 'Regressor', 'StochasticRegressor'
+    'GaussianCNNRegressorModel', 'GaussianMLPRegressor', 'Regressor',
+    'StochasticRegressor'
 ]
